Re-write the interface of the stm_builtin module.
Estimated hours taken: 1.5
Branches: main

Re-write the interface of the stm_builtin module.  This change is effectively
half of my review comments concerning Leon's previously posted diff.

XXX most of the C code in this module will be moved into the runtime as part
of a separate change.  I've left it largely untouched in this change.

library/stm_builtin.m:
	Fix the copyright message; we didn't start working on this
	file in 2004.

	Rename the tvar/1 type to stm_var/1.  (The former name is used
	extensively in the Mercury compiler itself, and is not particularly
	meaningful in any case.)

	Define a type to represent whether a transaction is valid or not.
	Use this instead of the bool/0 type.  (This avoids us having
	to always implicitly import the bool module when atomic scopes
	are used.)

	Document the predicates and types in this module more extensively.

	Conform to the above changes.

	Add some XXX comments about things that need to fixed.