Add a higher-order predicate for atomically executing closures using
Estimated hours taken: 2
Branches: main

Add a higher-order predicate for atomically executing closures using
STM.  This will eventually be replaced by the atomic scopes, but until
they are available having this facility available assists with working
on the STM runtime support.

This diff also includes the Mercury interface to the STM retry operation.
(I was working on this when I realised the above would be useful, which
is why it is in the same workspace.)  The runtime support for retry is
not included in this diff.

library/stm_builtin.m:
	Add a higher-order predicate that atomically executes a closure
	using STM.  (This is like the one in the prototype STM implementation
	except that it handles exceptions properly.)

	Add the Mercury interface to the STM retry operation.
	(The runtime support for this has not been implemented; it will
	currently cause execution to abort.)

runtime/mercury_stm.h:
runtime/mercury_stm.c:
	Add a stub function for MR_stm_retry_impl.