Fix a bug in the bug in the generation of if-then-else goals in the Erlang
Estimated hours taken: 2
Branches: main

Fix a bug in the bug in the generation of if-then-else goals in the Erlang
backend.  The calculation of the variables bound by the then branch was wrong
for model_semi conditions -- it included variables bound in the condition.

compiler/erl_code_gen.m:
	Fix the calculation of variables bound by then branches.

tests/valid/Mmakefile:
tests/valid/erl_ite_vars.m:
	Add test case.