Address bug #24.  The debugger help message referred to the `fail' command,
Estimated hours taken: 0.5
Branches: main

Address bug #24.  The debugger help message referred to the `fail' command,
even though the documentation for it is commented out in the user's guide.
The command had been left out of the table of commands in
the trace directory, so the debugger didn't respond to it anyway.

Since the command doesn't work properly, which is why the documentation
in the user's guide is commented out, remove all other user visible
documentation that refers to it.

Add the `fail' command to the debugger's internal table of commands so
that developers can use it.

doc/mdb_categories:
	Delete `fail' from the forward execution commands.  It doesn't
	work properly yet.  (See the comments in doc/user_guide.texi.)

trace/mercury_trace_internal.c:
	Add the `fail' command to the table of debugger commands.