Always compile the programs in the deep_profiler directory.
Estimated hours taken: 0.1
Branches: main

Always compile the programs in the deep_profiler directory.  Given the
increased dependence of this code on the mdbcomp directory it is sensible to
always compile the code, not just when the deep profiler is enabled.

Installation of the deep profiler is still conditional on the existence of
the CGI lib directory etc.

Mmakefile:
deep_profiler/Mmakefile:
	Always build the dependencies and compile the deep_profiler directory.