Add support for filtering the set of library grades to be installed
Estimated hours taken: 4
Branches: main

Add support for filtering the set of library grades to be installed
from the command line.  This is useful, for example, if you wish to
install a library in all the trailing grades that are installed with
a particular system.  (The new options are only useful with `--make',
as the library grade set is handled externally of the compiler with Mmake.)

This change also adds some sanity checking of the library grade set
immediately after the invocation of the compiler.  Previously this
was only done at point when a library was being built in a particular
grade.  (The sanity checking is not complete, since it is not currently
possible to check whether a grade is valid or not without modifying
the option table -- something this change cannot do.)

compiler/options.m:
	Add two new options, `--libgrades-exclude-component' and
	`--libgrades-include-component', that allow the set of library
	grades to be installed to be filtered by component from the
	command line.

compiler/handle_options.m:
	Add code to post-process the library grade set and apply any
	grade filters to that set.

	Break some overlong lines.

	Unrelated change: fix an XXX regarding the Erlang backend.

doc/user_guide/texi:
	Document the new options.

	Fix a typo: s/segement/segment/.