Fix a bug introduced with the last ssdb change to ml.in.
Estimated hours taken: 4
Branches: main

Fix a bug introduced with the last ssdb change to ml.in. The context of the bug
was the renaming of a bunch of sh variables from TRACE_xxx to TRACE_BASE_xxx
(for several values of xxx). The bug itself was that the defining occurrence
of one such variable (TRACE_LIBS_SYSTEM) was renamed, but a reference to it
(in the true arm of the switch of $trace_base) was not. The result was a
reference to an undefined variable, which the shell replaced with "".
Since $TRACE_LIBS_SYSTEM should have contained -ldl on some systems (including
my laptop), the upshot was that all attempts to link debuggable executables
failed, with the linker unable to find the definitions of the functions defined
in libdl.a.

scripts/ml.in:
	Fix that bug.

	Change to four-space indentation in order to avoid bad line breaks,
	though using tabs (for speed of interpretation) rather than spaces.

	Fix deviations from good sh programming style, including missing ;;s
	at the end of case arms.

configure.in:
	Rename TRACE_LIBS_SYSTEM to TRACE_BASE_LIBS_SYSTEM, since having the
	TRACE_BASE_LIBS_SYSTEM sh variable in ml.in get its value from a
	configure variable named TRACE_LIBS_SYSTEM is confusing (since ml.in
	also has a sh variable named TRACE_LIBS_SYSTEM).

scripts/ml.in:
scripts/Mercury.config.in:
scripts/Mercury.config.bootstrap.in:
	Conform to the change to configure.in.