Fix one of the problems that causes the compiler to run out of stack
Estimated hours taken: 2
Branches: main

Fix one of the problems that causes the compiler to run out of stack
space when compiling a discriminated unions that has a large numbers of
constructors.  This speeds up compilation of a type with 200000 constructors
by about 1% -- it doesn't have a measurable impact on discriminated unions
of a more usual size.

compiler/prog_io.m:
	When checking that the constructors of a discriminated union
	are well-formed, iterate through the list of constructors once instead
	of four times.