Fix a problem with the MLDS backend that caused it to generate invalid C
Estimated hours taken: 6
Branches: main

Fix a problem with the MLDS backend that caused it to generate invalid C
for static ground terms with no tag types with float arguments.
This is bug #37 in Mantis.

The problem was caused by the code in ml_unify_gen.m that generates
initializers for static terms calling a predicate that instead generates
code for constructing terms on the heap.  This "works" in most cases
except for with floats where we need to call MR_box_float() if we are
constructing them on the heap.  This fix is to call a predicate
that generates C that is valid in a static initializer.

compiler/ml_unify_gen.m:
	Fix an incorrect predicate call that lead to invalid
	initializers for static data being generated.

	Pass some extra arguments that are required by
	the correct predicate around.

compiler/ml_code_util.m:
	Fix the wording of a comment.

tests/hard_coded/Mmakefile:
tests/hard_coded/static_no_tag.{m,exp}:
	Regression test for the above bug.
	Also, exercise the construction of static ground terms
	containing no tag types a bit more.