1) Implement the saving of protected regions at disj frames ONLY for semidet
Estimated hours taken: 8
Branch: main

1) Implement the saving of protected regions at disj frames ONLY for semidet
disjunctions. We need this information so as to be able to commit the regions'
removals when a nonlast disjunct of a semidet disjunction succeeds.

Now the fixed part of a disj frame will have 4 slots instead of 3. The extra
slot is to store the number of protected regions. It will be zero for nondet
disjunctions.

For semidet disjunctions we can use the information in their rbmm_goal_info
for protected regions and allocated-into regions.

Also implement the profiling and debugging messages for the saving.

2) Fix a bug relates to the saving of region size records at ite frame.

3) Do not protect regions that are explicitly removed at the end of the
condition of an if-then-else. It is because when the program reaches that
point, we know for sure that the condition will succeed.

4) Improving the profiling of RBMM

compiler/disj_gen.m:
	Implement 1).
	Remove the old code that do the saving for nondet disjunctions.

compiler/llds.m
compiler/llds_out.m
compiler/opt_debug.m
	Make them conform to the changes for 1).

compiler/options.m
	Make sure the size of the fixed part of disj frames is 4 (slots).
	Change to conform to the changes for 1).

runtime/mercury_region.h
runtime/mercury_region.c
	Implement the saving and its profiling and debugging messages.
	Implement the improvement of the profiling of RBMM:
		more information,
		more structured when displaying the profiling information,
		correct some bugs.

runtime/mercury_types.h:
	To conform to the changes at 1).