Add a new option to configure, --with-llds-base-grade=..., that overrides the
Estimated hours taken: 1
Branches: main

Add a new option to configure, --with-llds-base-grade=..., that overrides the
configure script's choice of LLDS base grade.  This is useful when the
configure script detects the presence of some gcc extensions, e.g. global
registers, but Mercury cannot be compiled because of a buggy version of gcc.
Unlike specifying the grade in an Mmake.params file, specifying the base grade
to configure allows the set of library grades to be picked appropriately.
It's also easier to tell users who have buggy versions of gcc to just do,
e.g.  --with-llds-base-grade=none, rather than work out what all the required
workarounds for their particular system may be.

configure.in:
	Add the new option.