In the handling of `:- pragma type_spec' declarations, a proc_info may be
Branches: main

In the handling of `:- pragma type_spec' declarations, a proc_info may be
derived from a proc_info that holds structure sharing or reuse information.
Since the specialised procedure won't have the same interface as the original
procedure, that information won't be directly usable so we need to delete it.

We do a similar thing in higher_order.m.  However, when making optimisation
interfaces the higher order specialisation pass won't be run so we do it in
add_pragma.m as well.

compiler/add_pragma.m:
	As above.