Avoid assertion failures in the liveness detection pass which is called as
Estimated hours taken: 1
Branches: main

Avoid assertion failures in the liveness detection pass which is called as
part of computing in-use information for structure reuse.  The problem occurs
for procedures containing an if-then-else goal with an `erroneous' condition.
As written, liveness.m expects such goals to have been simplified beforehand.
This is not the case when structure reuse is enabled, as the structure reuse
analysis is run before simplification.

compiler/structure_sharing.analysis.m:
	Fix the problem by simplifying procedures before liveness detection.

compiler/liveness.m:
	Add a note about expecting simplified goals.