Allow the use of munmap() in Boehm GC, in order to return unused memory pages
Estimated hours taken: 6
Branches: main

Allow the use of munmap() in Boehm GC, in order to return unused memory pages
back to the OS.  This has an adverse effect on run times so we want to be able
to enable it on a per-application basis.  To avoid installing both
munmap-enabled and munmap-disabled copies of Boehm GC, we make it so that
munmap() capability can be compiled in but disabled with a global variable.
In that case the performance difference (seems to) be negligible.

As mmap/munmap aren't supported on all platforms, munmap() capability needs to
be selected with `configure'.  It is currently disabled by default.

Mmake.common.in:
configure.in:
	Add the configure option `--enable-gc-munmap' and
	add -DUSE_MMAP -DUSE_MUNMAP to the Boehm GC C flags if set.

boehm_gc/allchblk.c:
boehm_gc/alloc.c:
boehm_gc/malloc.c:
boehm_gc/include/gc.h:
boehm_gc/include/private/gc_priv.h:
	Add the global variable `GC_mercury_use_munmap'.

	Modify the code to check `GC_mercury_use_munmap' before executing code
	that would normally be run when USE_MUNMAP is set.

runtime/mercury_wrapper.c:
	Set `GC_mercury_use_munmap' if MERCURY_OPTIONS contains `--munmap'.

doc/user_guide.texi:
	Document the new MERCURY_OPTIONS and configure options.