MERCURY_OPTIONS is a crude way to give options to Mercury programs, since it
Estimated hours taken: 1
Branches: main

MERCURY_OPTIONS is a crude way to give options to Mercury programs, since it
applies to all Mercury programs, even though we often want to give options to
only one specific Mercury program. (For example, we may want to give it only
to a test program's executable, and not to the invocation of the compiler
that generates it.) This diff makes the runtime system, when executing a
program named "progname", look for and process the environment variable
whose name is "MERCURY_OPTIONS-progname". This allows options to be given
at runtime to only one specific program.

runtime/mercury_wrapper.c:
	Make the change described above.

runtime/mercury_wrapper.h:
	Fix some obsolete documentation, and add some up-to-date documentation.

doc/user_guide.texi:
	Document the change.

util/mkinit.c:
	Fix indentation.