Wrap the definitions of the standard library's stack/1 and bag/1 types in
Estimated hours taken: 0.5
Branches: main

Wrap the definitions of the standard library's stack/1 and bag/1 types in
notag wrappers.  This prevents problems with overlapping instances when they
are used in type class instance definitions and in the case of bag/1 makes it
usable in such definitions (the current definitions does not conform to
what is allowed in instance heads under the current system.)

This change does mean that the source code for these modules is a little
more verbose, but this is outweighed by the fact that the behaviour with
instances is much less surprising than it was.

library/bag.m:
library/stack.m:
	Define these types using notag wrappers rather than equivalence types.