Add a soft limit on the number of structure reuse conditions that we will
Branches: main

Add a soft limit on the number of structure reuse conditions that we will
accumulate for any given procedure.  Too many reuse opportunities on a
procedure are useless if no callers will be able to satisfy the reuse
conditions on the procedure.  We don't know which conditions are unlikely to be
satisfied, but the number of conditions ought to be a decent approximation.
Also, large sets of reuse conditions take more space and time to work with
(e.g. parsing, verifying).

compiler/options.m:
	Add the developer-only option `--structure-reuse-max-conditions <n>'.

compiler/structure_reuse.domain.m:
	Add a function to count reuse conditions.

compiler/structure_reuse.direct.choose_reuse.m:
	If a found direct reuse opportunity would make the number of reuse
	conditions exceed the limit, then ignore that reuse opportunity.

compiler/structure_reuse.indirect.m:
	If the number of reuse conditions is too high, don't make any more
	calls to reuse procedures which have conditions on them.