Avoid a linking problem that can occur when using structure reuse and
Branches: main

Avoid a linking problem that can occur when using structure reuse and
`--intermodule-analysis'.  When making an `.analysis' we may find that a
procedure has conditional reuse, but when making the target code we may find
that procedure has unconditional or no reuse.  If another module calls the
conditional reuse version of the procedure, which doesn't exist, that would
lead to a linking error.

The solution is to produce a fake conditional reuse procedure which forwards
to the unconditional/no-reuse procedure if we detect this situation.  We
already did that when using `--intermodule-optimisation'.

compiler/structure_reuse.analysis.m:
	As above.