Add a new FAQ section 'installation', and create two questions within it.
Estimated hours taken: 0.5
Branches: main

Add a new FAQ section 'installation', and create two questions within it.

The first question describes a work-around for bugs 66 and 77.

The second question describes why the make install process takes a while and
directs the user to the Grades section in the User Guide.

doc/faq.texi:
	As above.