Implement a new option --allow-some-paths-only-waits, that tells dep_par_conj.m
Estimated hours taken: 32
Branches: main

Implement a new option --allow-some-paths-only-waits, that tells dep_par_conj.m
that it is ok to insert a wait for a variable along some but not all execution
paths in a branched control structure such as a switch or if-then-else. If this
happens, then we need to wait separate for any uses of the variable in later
code, and we rename the consumed variable in the branched control structure
to avoid defining it twice.

The effect of the option is shown by this partial HLDS dump from after
the dep_par_conj transformation for the consume_in_some_branches_and_after
test case:

     (if
       consume_in_some_branches_and_after.test(Y_5)
     then
-      A_7 = Y_5
+      ( % conjunction
+        A_7 = Y_5
+        ,
+        par_builtin.wait(FutureX_8, X_4)
+      )
     else
       ( % conjunction
-        par_builtin.wait(FutureX_8, X_9)
+        par_builtin.wait(FutureX_8, X_4)
         ,
-        A_7 = X_9
+        A_7 = X_4
       )
     )
     ,
-    par_builtin.wait(FutureX_8, X_4)
-    ,
     Z_6 = int.(X_4 + A_7)

The lines starting with + are with --no-allow-some-paths-only-waits, while
the lines starting with - are with --allow-some-paths-only-waits.

The advantage of --allow-some-paths-only-waits is that if execution takes
a path through the branched control structure in which the consumed variable
is NOT needed, then we can delay the wait until the next consumer of the
variable, which may be substantially later. The disadvantage is that if
execution takes a path through the branched control structure in which the
consumed variable IS needed, then we will wait for it more than once.

The disadvantage is a small constant factor while the advantage may be
substantial. In absence of further analysis (which could estimate the size
of both the advantage and the disadvantage), our best guess is that the
advantage wins.

compiler/options.m:
	Add the new option.

compiler/dep_par_conj.m:
	Implement the new option.

tests/par_conj/consume_in_some_branches_and_after.m:
	Replace a unification with a call to allow the procedure containing
	that code to serve as a simple test case of the option (the one
	shown above). The computation being described remains unaffected.