Java backend name mangling fun:
Branches: main

Java backend name mangling fun:

1. Java doesn't allow a class name to have the same name as a package, which
conflicts with Mercury nested modules.  Work around this problem by suffixing
package name components with underscores.  For example, the Mercury module
`mammal.primate.chimp' would correspond to a Java class `chimp' in the package
`mammal_.primate_'.  The exceptions to this are the `mercury' and
`mercury.runtime' as the first two package name components.

2. When compiling the standard library, javac aborts on the name of a `.class'
file generated for a nested class, saying the file name is too long.  Avoid the
problem by arbitrarily shortening the names of wrapper classes generated for
taking the addresses of predicates.  (In the future we may need to do more of
the same in other contexts but this is enough for the standard library.)

compiler/mlds_to_java.m:
        As above.

compiler/java_util.m:
        Add functions to return the package names used by the Java backend.