Avoid a problem with type names and constructor names that are the same.
Branches: main

Avoid a problem with type names and constructor names that are the same.
Java does not allow the name of a nested class to be the same as its
enclosing class.  If we would violate that rule (when generating Java) then
add the prefix "mr_" to the constructor name.  The user can still trigger the
problem by naming a different constructor with the "mr_" prefix.

compiler/ml_type_gen.m:
        Add functions to generate constructor names that avoid the problem.

        Use the functions for naming d.u. constructors.

compiler/ml_unify_gen.m:
        Use the functions for naming d.u. constructors.

compiler/mlds_to_java.m:
        Delete TODO.