Use java.lang.Object[] as the type of all non-primitive arrays instead of more
Branches: main

Use java.lang.Object[] as the type of all non-primitive arrays instead of more
specific types.  To create arrays with the more specific types we used
reflection on a representative element, but we can't do that if the element is
`null', as it might be for foreign types.

compiler/mlds_to_java.m:
        Write java.lang.Object[] instead of Type[] when Type is not a primitive
        type.

library/array.m:
        For non-primitive element types, just create an Object[] array.