Fix a performance problem with mmc --make --jobs <n>.
Branches: main

Fix a performance problem with mmc --make --jobs <n>.  It would fork a new
process for each target, whether or not that target was already up-to-date.
For a large program which is mostly up-to-date, this could be very costly.
The main reason it was implemented so was to avoid interprocess communication
which might be non-portable (the only communication was via a sub-process's
exit code).

This patch makes use of IPC to allow the master process to fork off only N-1
worker processes, itself being the last worker.  The workers communicate via
shared memory as to which tasks still need doing, and which are completed.
A worker process therefore may work on as many tasks as required before
terminating.

Parallel mmc --make is currently disabled in .par grades.  It should be
fairly simple to restore.


configure.in:
runtime/mercury_conf.h.in:
        Check sys/mman.h is available for mmap().

Mmake.common.in:
compiler/Mmakefile:
        Link with thread libraries for POSIX mutexes.

compiler/make.util.m:
        Reimplement process-based concurrent fold predicate as above.
        Fall back to non-concurrent fold if the IPC mechanisms are
        unavailable.

        Delete old thread-based concurrent fold implementation.

compiler/process_util.m:
        Add wait_pid to wait on a particular child process.

compiler/Mercury.options:
        Set --no-ansi-c on make.util.m.