Make version_array.set make a new array if it's not updating the latest version
Branches: main

Make version_array.set make a new array if it's not updating the latest version
of an array.  This avoids unexpected bad performance, where the version array
essentially turns into a linked list with an array at the end, if the user
happens to go back to an older version of an array without an explicit
rewind/copy.

library/version_array.m:
        As above.

compiler/make.dependencies.m:
        Use size doubling when resizing version arrays.