Parallel runtime profiling corrections.
This patch makes some corrections to my previous patch.

I've fixed some printf format strings so that they are correct on 32bit and
64bit platforms.  Some fields in the parallel_execution_profile.txt file so
that their meaning is move obvious.  Fixed a stray 'i' that would have caused
a syntax error, this would have only been noticed on Taura's nightly builds
since this code is not included in normal builds.  Corrected how runtime of
the MR_do_runnext() loop is measured, previously it included the time that an
engine was sleeping waiting for work.

runtime/mercury_context.c:
	As above.